Create java codes
Suppose that you want to write a program to calculate the marks of a student of all his subjects that he has learnt in a particular semester. Take the input from the user which will ask him to enter no. of subjects he has learnt. If the input given is greater than 0, then program should ask the user to enter the marks of each subject. After entering all the marks, calculate and print total marks scored by him in all subjects. If the input given is negative or zero, then display the message “Invalid”.